Update imports to reflect the renaming of dir from 'lib' to 'utils'
